feat: 영화 정보 리코일 전역 관리 설정
